bug update

+++ src/main/java/com/xieke/admin/dto/ResultInfo.java
@@ -15,7 +15,7 @@ public class ResultInfo<T> implements Serializable {
         this.data = data;
     }
 
-    public ResultInfo(T data, int count) {
+    public ResultInfo(T data, long count) {
         this.data = data;
         this.count = count;
     }
@@ -51,7 +51,7 @@ public class ResultInfo<T> implements Serializable {
     /**
      * 总记录数
      */
-    private int count;
+    private long count;
 
     public String getMsg() {
         return msg;
@@ -77,7 +77,7 @@ public class ResultInfo<T> implements Serializable {
         this.data = data;
     }
 
-    public int getCount() {
+    public Long getCount() {
         return count;
     }
